feat: Enable fractional share trading end-to-end
## Backend Changes

### Trading Config (src/trading/config.py)
- Added FractionalConfig with ALLOW_FRACTIONAL, MIN_FRACTIONAL_QTY,
  FRACTIONAL_PRECISION, USE_NOTIONAL_ORDERS settings
- Default: fractional trading enabled for paper trading

### Position Sizing (src/trading/automated_trading.py)
- _calculate_position_size() now returns float instead of int
- Respects ALLOW_FRACTIONAL config
- Rounds to 4 decimal places (Alpaca's precision)

### Portfolio Manager (src/agents/portfolio_manager.py)
- PortfolioDecision.quantity changed from int to float
- compute_allowed_actions uses float for share counts

### Alpaca Service (src/trading/alpaca_service.py)
- execute_decision() accepts float quantity
- submit_order() handles fractional orders:
  - Converts to market/DAY orders (Alpaca requirement)
  - Auto-retries with whole shares if asset not fractionable
  - Logs warnings when rounding occurs

### Database Models
- pm_decisions.quantity: Integer → Float
- trade_decision_context.pm_quantity: Integer → Float

## Frontend Changes

### Utility Functions (lib/utils.ts)
- formatQuantity(): Shows up to 4 decimals for fractional shares
- formatShares(): Adds "shares" suffix

### Components Updated
- TradingDashboard.tsx: Uses formatShares() for positions
- CommandCenter.tsx: Uses formatShares()/formatQuantity()
- RoundTable.tsx: Uses formatShares() for PM quantity
- investment-report-dialog.tsx: Uses formatQuantity()

## Configuration

Environment variables (all optional):
- ALLOW_FRACTIONAL=true (default)
- MIN_FRACTIONAL_QTY=0.0001
- FRACTIONAL_PRECISION=4
- USE_NOTIONAL_ORDERS=false

## Backwards Compatibility

Set ALLOW_FRACTIONAL=false to use whole shares only.
Non-fractionable assets automatically round to whole shares.

+++ app/backend/database/models.py
@@ -394,7 +394,7 @@ class TradeDecisionContext(Base):
     
     # PM decision
     pm_action = Column(String(20), nullable=True)  # buy, sell, short, cover, hold
-    pm_quantity = Column(Integer, nullable=True)
+    pm_quantity = Column(Float, nullable=True)  # Supports fractional shares
     pm_confidence = Column(Float, nullable=True)
     pm_reasoning = Column(Text, nullable=True)
     pm_stop_loss_pct = Column(Float, nullable=True)
